Add command to overwrite playlist.
MPD, tragically, does not have good functionality for updating/editing
an existing saved playlist. To update a playlist, what is actually done
is that the playlist is deleted and then recreated with the edits.

I originally tried to rework the `:w` and `:write` to allow for updating
an existing playlist. First I tried to always delete the playlist before
creating it (see #40). However, if you wanted to make a new playlist,
this removal would error, confusing the user. I wanted to only remove an
existing playlist if it existed but could not, for the life of me could
not figure out a way to do that.

Hopefully this works for now.